DEVONthink 3 - Smart Rename Apple Script
Apple Script to use smartrename.pl in the context of DEVONthink 3. It requires Perl (plus some Perl packages) to be installed and smartrename.pl to be present in a specific location. You need to change this to your local path to make it work. Obviously it should be deployed in a better way, but I didn't take time to figure that part out.